simcut/requirements.txt
altgraph==0.17.5
coverage==7.10.7
exceptiongroup==1.3.1
importlib_metadata==8.7.1
iniconfig==2.1.0
macholib==1.16.4
packaging==26.0
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.1
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.10.2
pytest==8.4.2
pytest-cov==7.0.0
pytest-qt==4.5.0
tomli==2.4.0
typing_extensions==4.15.0
zipp==3.23.0
